chore: 패키지명 변경 및 적용
