+++ src/main/java/com/waggle/domain/projectV2/controller/ProjectV2Controller.java
@@ -26,7 +26,7 @@ import com.waggle.global.response.swagger.MembersSuccessResponse;
 import com.waggle.global.response.swagger.ProjectV2SuccessResponse;
 import com.waggle.global.response.swagger.ProjectV2sSuccessResponse;
 import com.waggle.global.response.swagger.RecruitmentsSuccessResponse;
-import com.waggle.global.secure.oauth2.CustomUserDetails;
+import com.waggle.global.security.oauth2.CustomUserDetails;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
 import io.swagger.v3.oas.annotations.media.Schema;
